+++ NEWS.md
@@ -1,6 +1,7 @@
 # fmtr 1.7.0
 
 * Added `fmt_mean_stderr()` helper function.
+* Handle variations on "best" format specification.
 
 # fmtr 1.6.9