+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Added `fmt_mean_stderr()` helper function.
 * Handle variations on "best" format specification.
+* Added "dateW" formats to `fapply()`.
 
 # fmtr 1.6.9