+++ NEWS.md
@@ -1,3 +1,8 @@
+# fmtr 1.7.1
+
+* Added "timeW.d" formats to `fapply()`.
+
+
 # fmtr 1.7.0
 
 * Added `fmt_mean_stderr()` helper function.